Restrict image downloads to Discord CDN URLs (#80)
Added a check to ensure that only images hosted on Discord's CDN are downloaded and verified on the Bot API. This improves security by preventing downloads from arbitrary URLs.

This is not really necessary as the URL is provided by the bots from a trusted source but it shuts up the code scanners.